stormshield/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1

# Streamlit + UI
streamlit==1.41.1
streamlit-folium==0.26.2
folium==0.19.4
plotly==5.24.1

# Data & ML
pandas==2.2.3
numpy==2.2.1
scipy==1.15.1
xgboost==2.1.3
scikit-learn==1.6.1
joblib==1.4.2

# Pydantic & Settings
pydantic==2.10.5
pydantic-settings==2.7.1

# Scheduler
apscheduler==3.11.0

# Google Gemini
google-generativeai==0.8.4

# RSS / scraping
feedparser==6.0.11
beautifulsoup4==4.12.3
lxml==5.3.0

# Logging & utils
python-dotenv==1.0.1
